+++ packages/nimble-components/src/select/styles.ts
@@ -78,7 +78,7 @@ export const styles = css`
     }
 
     .filter-field:before {
-        bottom: calc(${controlHeight} + ${smallPadding} + (2 * ${borderWidth}));
+        bottom: calc(${controlHeight} + ${smallPadding} - ${borderWidth});
     }
 
     .filter-field:not(.above)::after,
@@ -113,14 +113,7 @@ export const styles = css`
         height: ${controlHeight};
         display: inline-flex;
         align-items: center;
-    }
-
-    .no-results-label.below {
-        padding: ${smallPadding} ${mediumPadding} 0 ${mediumPadding};
-    }
-
-    .no-results-label.above {
-        padding: 0 ${mediumPadding} ${smallPadding} ${mediumPadding};
+        padding: ${smallPadding} ${mediumPadding};
     }
 `.withBehaviors(
     appearanceBehavior(